Refactor rccl comm data collection
Mode tool_tracing_callback_rccl to it's own file.
Removing some debug logs
