+++ source/lib/rocprof-sys/library/components/comm_data.hpp
@@ -77,18 +77,6 @@ struct comm_data : base<comm_data, void>
         static constexpr auto label = "MPI Comm Send";
     };
 
-    struct rccl_recv
-    {
-        static constexpr auto value = "comm_data";
-        static constexpr auto label = "RCCL Comm Recv";
-    };
-
-    struct rccl_send
-    {
-        static constexpr auto value = "comm_data";
-        static constexpr auto label = "RCCL Comm Send";
-    };
-
     ROCPROFSYS_DEFAULT_OBJECT(comm_data)
 
     static void preinit();
@@ -145,55 +133,6 @@ struct comm_data : base<comm_data, void>
     static void audit(const gotcha_data& _data, audit::incoming, const void*,
                       int sendcount, MPI_Datatype sendtype, void*, int recvcount,
                       MPI_Datatype recvtype, MPI_Comm);
-#endif
-
-#if defined(ROCPROFSYS_USE_RCCL)
-    static auto rccl_type_size(ncclDataType_t datatype)
-    {
-        switch(datatype)
-        {
-            case ncclInt8:
-            case ncclUint8: return 1;
-            case ncclFloat16: return 2;
-            case ncclInt32:
-            case ncclUint32:
-            case ncclFloat32: return 4;
-            case ncclInt64:
-            case ncclUint64:
-            case ncclFloat64: return 8;
-            default: return 0;
-        };
-    }
-
-    // ncclReduce
-    static void audit(const gotcha_data& _data, audit::incoming, const void*, const void*,
-                      size_t count, ncclDataType_t datatype, ncclRedOp_t, int root,
-                      ncclComm_t, hipStream_t);
-
-    // ncclSend
-    // ncclGather
-    // ncclBcast
-    // ncclRecv
-    static void audit(const gotcha_data& _data, audit::incoming, const void*,
-                      size_t count, ncclDataType_t datatype, int peer, ncclComm_t,
-                      hipStream_t);
-
-    // ncclBroadcast
-    static void audit(const gotcha_data& _data, audit::incoming, const void*, const void*,
-                      size_t count, ncclDataType_t datatype, int root, ncclComm_t,
-                      hipStream_t);
-
-    // ncclAllReduce
-    // ncclReduceScatter
-    static void audit(const gotcha_data& _data, audit::incoming, const void*, const void*,
-                      size_t count, ncclDataType_t datatype, ncclRedOp_t, ncclComm_t,
-                      hipStream_t);
-
-    // ncclAllGather
-    // ncclAlltoAll
-    static void audit(const gotcha_data& _data, audit::incoming, const void*, const void*,
-                      size_t count, ncclDataType_t datatype, ncclComm_t, hipStream_t);
-
 #endif
 
 private:
